backend/pyproject.toml
[tool.poetry]
name = "expenditures"
version = "0.1.0"
description = ""
authors = ["Daniel Grießhaber <griesshaber@hdm-stuttgart.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-Cors = "^3.0.10"
flask-marshmallow = "^0.14.0"
flask-sqlalchemy = "^3.0.2"
Flask-BasicAuth = "^0.2.0"
marshmallow-sqlalchemy = "^0.28.1"

[tool.poetry.dev-dependencies]
mypy = "^0.982"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
